+++ springboot/src/main/java/com/mingde/service/NoticeService.java
@@ -5,34 +5,36 @@ import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.mingde.entity.Notice;
 import com.mingde.mapper.NoticeMapper;
+import com.mingde.utils.AuthUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-/**
- * 业务层方法
- */
 @Service
-public class NoticeService{
+public class NoticeService {
 
     @Autowired
     private NoticeMapper noticeMapper;
 
     public void add(Notice notice) {
+        AuthUtils.requireAdmin();
         notice.setTime(DateUtil.now());
         noticeMapper.insert(notice);
     }
 
     public void updateById(Notice notice) {
+        AuthUtils.requireAdmin();
         noticeMapper.updateById(notice);
     }
 
     public void deleteById(Integer id) {
+        AuthUtils.requireAdmin();
         noticeMapper.deleteById(id);
     }
 
     public void deleteBatch(List<Integer> ids) {
+        AuthUtils.requireAdmin();
         for (Integer id : ids) {
             noticeMapper.deleteById(id);
         }
@@ -47,9 +49,9 @@ public class NoticeService{
     }
 
     public PageInfo<Notice> selectPage(Notice notice, Integer pageNum, Integer pageSize) {
-        PageHelper.startPage(pageNum, pageSize);
+        AuthUtils.requireAdmin();
+        PageHelper.startPage(pageNum, Math.min(pageSize, 100));
         List<Notice> list = noticeMapper.selectAll(notice);
         return PageInfo.of(list);
     }
-
 }
